// FE/rempms-candidate-frontend/src/app/model/candidate/familyInformation/FamilyInformation.ts
export class FamilyInformation {
  idFamilyInformation: string | any;
  name: string | any;
  relationship: string | any;
  dob: string | any;
  gender: string | any; // enum : gender
  designation: string | any;
  schoolOrOrganization: string | any;
  remark: string | any;
  isDependant: boolean | any;
}